OPEN and SHUT can move into each other even when STOP
Slight issue with this implementation: if the key or door is on a tile for example, you can't push it in.
For pushing, we can't just run try_move on everything. Instead we need to check if they can move, then only push everything if all pass. Likewise, we should be filtering out the corresponding property from the STOP check instead of having an additional overriding check