Time to redo text parsing yet again!
First I tried to fix a bug where 'and A is B and' would highlight the ands, then I realized that 'A is B and has C' is a valid sentence, meaning verbs can be sequenced. This of course throws out my entire assumption that there was only one verb per sentence, so we could start parsing there. Instead, we need to start parsing at every word that doesn't have another word to its left. This actually causes a bug in the vanilla game where a sentence like 'not A is B' with another random piece of text stacked on the 'not' will actually form both 'not A is B' and 'A is B.' So anyway, let's do it.