Work in progress attempt: one function per state?
I'm really struggling with the direction on this. We need to parse text from left to right, where every subsequent piece of text splits everything so far in case of text stacking. Therefore, we need to cache the rules in the entities themselves, since storage will be clobbered by the time "as" moves on. I have three main ideas:
1. State machine with one function per state. I made a 16-state NFA for the Baba text parser, where each state was the same as all transitions to it. However two bits of nondeterminism complicate it slightly: there are two different 'not' start states, and the 'not' after an 'and' after an infix could either be before an infix or a noun.
2. Happy path parser for all alternatives. At the start of the string, it would try to parse both a prefixed subject, and a non-prefixed one, right after the other. Both would have separate but similar steps, moving the context down the line.
3. Bring back the teleport parser. This would help with some shared constructions, like "not* X (and not* X)?" which appears several times, even nested. I feel like it has unforeseen complications.