Improve renderer
Go column by column instead of row by row. If we try to draw an object that's next to something with a lower Z layer, do that one first. If it's on an adjacent column, we have to insert spaces before and after. Also, finally replaced advances with splitting spaces in key places to prevent clipping. Also the FLOAT property now affects Z layer.