Level data fixers
As I build levels in the world, a major challenge is the noteblock IDs getting mixed up, or the jukebox level format changing. The former requires me to unpack all the levels to fix the blocks, the latter requires me to repack all the levels to fix the jukebox. If I change both, the levels are hard to recover.
So now I have a thing that will go through and repack all of the levels I have built. Some things like the Z coordinate and X offset are hardcoded, but it handles additional levels just fine. Loading the chunks as it goes is accomplished with the /clone hack, and one level is saved per tick with /schedule, which is still fast without blowing the command limit.
Also some minor editor improvements, the most important one being unpacking now places gray concrete on the bottom layer to make building there easier