Optimization: parse a reduced subset of text
In an average step, not much changes. But currently we're still naively throwing away all rules, re-parsing them, and re-assigning them to objects.
With this optimization, we mark certain text in need of reparsing: any text that is created or moved, or any text next to some text that was destroyed or moved. If none of this happened in a step, we can freely skip text updating, which is a huge save. If there are some, we could fully reparse, but we can do a bit better. First we flood-fill any text marked to reparse, since any part of the sentence could be ultimately affected. We remove any rules that any of those text pieces participated in, then finally run parsing on just those.