Big sad realization
Assigning rules to objects, i.e. the more expensive part of rule updating, can't be skipped because of rules with conditions.
I also can't think of a way to only partially apply rules to objects, like removing the effects of removed rules and adding the effects of newly parsed rules. That's really tricky because the the method by which rules are applied basically leaves no trace of irrelevant rules. Like if you break an "A is not B" rule, entities don't know intrinsically whether they should now become B (i.e. if there is an "A is B" rule somewhere that applies to them) without checking again.
We already have a mechanism of only checking a subset of entities -- newly spawned ones after a transform. So in theory we could only check entities that moved, just like we do with text. OFTEN and SELDOM pose a problem with that idea, but to be fair I don't care very much about new adventures keywords. I thought about possibly caching these conditions in the entity itself so we can check them each step independently of rule application, but that doesn't work because constructs like "often A is not B" affect rule application.