Add ALL noun
Only works as a subject for now, but does rule disabling properly.
ALL as a condition or transform is weird -- it refers to every object ever mentioned in the level. So any sprite or noun text at level load, even if some gets destroyed later. For conditions, the subject must find a match for everything in this list, and for transforms, the subject turns into one of each.