Update palette generation
It's now hardcoded and direct instead of dynamic and recursive. This is a bit less pretty to me, but simplifies it a lot.
Since letters can't cache their part of speech with a tag, we need to be able to determine it dynamically by storing which words are which parts. Additionally, you can only spell words that are in the palette. These lists segregated by part can accomplish this.