Fix text disabling
Previously, we would iterate over all rules that could potentially disable, and find any rules that they did. However, the issue with that is that the same rule might get disabled multiple times: "flag is win and blue" should only cross out the "win" when a "flag is not win" rule exists, but previously, if more "flag is not win" rules existed, the "is" and "and" could get disabled too.
Now, we iterate over all rules that can be disabled, and find any rules that disable them. I think this way is no longer optimized for the common case of no disabling rules, but it can break early on success, and totally addresses the primary issue.
I haven't tested super thoroughly, but it seems to be correct.