Use marker entities for rules
This worked on the first try!
Instead of using a storage array for rules, we store them in marker entities. Of course, it's still storage during parsing, but gets converted during finalization. That has an upfront cost, but the save is that we can simply use selectors to iterate through rules instead of with recursion. Storage comparisons become comparatively cheap tag and score lookups.
Also, we can cache the rendered text form of the rule in the marker's CustomName, which is just lovely!
Of course, storage is fast, so it's worth measuring how much time this even saves, if anything.